investment_screening/requirements.txt
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0